htable.c: case insensitive keys and use list.h
Each element now duplicates the key data in lower case.
This also means that an update of the key data in the original position
(for example, a rename of a sector) no longer directly affects the
htable and will need a separate update.

Also use list.h instead of home-grown list.