server.c: disable accept for some time if overloaded
If a connection could not be accepted to due the process having too many
open files, libev unfortunately retriggers on the same connection
attempt in the next event loop iteration. This causes a continuous loop,
hogging all CPU in the server thread.

As of this commit, we stop the server accept watcher for 30 seconds when
this happens. This is bit of a dirty fix, but it lowers the system
impact considerably.