rbtree: use for neighbouring sector lookups
Organise all sectors in a red black tree sorted by their x coordinates
and use this tree to search for neighbouring sectors. This scales a lot
better than the old search.

A quirk with the new method is that all systems need to have unique x
coordinates, or there will be tree collisions. This commit also adds
that check to sector_move().