fix races between connection and server
The old way with message passing between the worker and server threads
was vulnerable to a race condition between the connection cleanup
handler and the worker threads. At any time, the connection struct could
be freed, even if a worker was currently active on it.

The new way moves connection cleanup to server.c which is called via an
asynchronous event using the libev loop in server. This avoids any
potential concurrency issues in libev as well.

A side effect of this change is that the message passing pipe between
the server and worker threads can now be removed, which is also done.