loadconfig: add and use for items, names, planets, ports and ships
Add the master configuration file yastg.conf and parsing of that file in
loadconfig. Use that code for loading items, names, planets, ports and
ships from the files specified in yastg.conf, instead of from hardcoded
paths relative to the current directory.

Configuration files are found using libxdg to try to keep in line with
modern standard. Add this library as a dependency in autotools.