config: remove parameters from load_all*() functions
These functions took the list head where the configuration data should
be put as a parameter. This made them general and versatile, but it
won't work when we move the configuration data to separate files.

In the future, this should be refactored, but it works and makes for
clean code right now.