names: enforce unique names in create_unique_name()
The name "create_unique_name" suggests it guarantees unique names, so
we'd better do that. Store all previously generated names in a string
tree at look them up. There's better ways, both in terms of cpu and
memory to fix this, bit this was the fastest.

This also does away with a name uniqueness-loop in port.c.