cli: disallow adding invalid commands
cli_add_cmd() never called the same trim_and_validate() function as
cli_run_cmd() did. It was therefore possible to add a command that could
subsequently not be run, which is probably a bad idea. Disallow it by
running the same input validation in cli_add_cmd() and in cli_run_cmd().