test: add cli_test
While there are tests for the stringtree implementation, there were no
tests for the cli abstraction layer on top of stringtree. As some bugs
had crept into this layer over time, this was evidently not enough.

Add some simple tests for the cli layer to avoid this in the future.