replace mprintf() with printf()
Previously, there has been several threads competing over writing to
stdout for debugging purposes. It was very convenient to have a printf()
which guaranteed line integrity, which was mprintf().

Such printouts have been phased out over time, and there is now no point
in having such a function in the code anymore.

Remove the function and replace all calls to mprintf() with calls to
printf().