server: remove erroneous close()
Commit d666e07c, "server.c: Allow multiple listen sockets", replaced the
sockfd socket file descriptor with a linked list of sockets.
Unfortunately, the variable itself and the corresponding call to close()
at the end of server_main() were left behind.

When the server was shut down, it therefore tried to close an invalid
filehandle. Remove the erroneous variable and call to close().